servers/requirements.txt
fastmcp
google-genai
python-dotenv